+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.2.0
+* Added support for linking multiple targets per SlotLink.
+	* SlotLink animations with the old data-model have to be migrated with one convenient button-press.\
+		This has to happen when SlotLink data is already set up.\
+		Nothing gets lost, all will work as before.
+
 ## v0.1.6
 * The initial setup for a new SlotLink now tries to determine the target automatically.
 * Added button to setup the links for all Slots, and determine the target automatically if possible.